refactor: Extract App's router into its own component
`MainAppNavigator` and `RootNavigator` were declared inside of `App`
functional component

This would produce unintended re-render, or worst, unintended re-mount
of the entire router if the `App` would have to re-render due to
changes on upper context

Extracting those component from `App` would stabilize the component
tree
